# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Revolut-inspired reference fintech platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with phone/email + passcode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ONBOARDING
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts for current user
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a new currency account/pocket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{accountId}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account details
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{accountId}/transactions:
    get:
      operationId: getAccountTransactions
      tags: [accounts]
      summary: Get transactions for an account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: currency
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPage"

  # CARDS
  /cards:
    get:
      operationId: listCards
      tags: [cards]
      summary: List all cards for current user
      responses:
        "200":
          description: List of cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
    post:
      operationId: issueCard
      tags: [cards]
      summary: Issue a new virtual or physical card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueCardRequest"
      responses:
        "201":
          description: Card issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
  /cards/{cardId}:
    get:
      operationId: getCard
      tags: [cards]
      summary: Get card details
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Card details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
    patch:
      operationId: updateCard
      tags: [cards]
      summary: Freeze/unfreeze or update card settings
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCardRequest"
      responses:
        "200":
          description: Updated card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"

  # PAYMENTS
...
  /fx/rates:
    get:
      operationId: getFxRates
      tags: [fx]
      summary: Get live exchange rates
      parameters:
        - name: base
          in: query
          required: true
          schema:
            type: string
            example: USD
        - name: targets
          in: query
          required: false
          schema:
            type: string
            description: Comma-separated currency codes
            example: EUR,GBP,JPY
      responses:
        "200":
          description: Exchange rates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FxRates"
  /fx/convert:
    post:
      operationId: convertCurrency
      tags: [fx]
      summary: Convert between currencies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FxConvertRequest"
      responses:
        "200":
          description: Conversion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FxConversionResult"
  /fx/history:
    get:
      operationId: getFxHistory
      tags: [fx]
      summary: Get historical exchange rates for a currency pair
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Historical rates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FxRatePoint"

  # NOTIFICATIONS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
      required: [error, message]
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        countryCode:
          type: string
        accountType:
          type: string
          enum: [personal, business]
        passcode:
          type: string
      required: [email, firstName, lastName, countryCode, accountType, passcode]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        passcode:
          type: string
      required: [email, passcode]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        sessionToken:
          type: string
      required: [user, sessionToken]
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        phone:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        countryCode:
          type: string
        accountType:
          type: string
          enum: [personal, business]
        kycStatus:
          type: string
          enum: [pending, in_review, approved, rejected, not_started]
        onboardingStatus:
          type: string
          enum: [not_started, in_progress, completed]
        createdAt:
          type: string
          format: date-time
      required: [id, email, firstName, lastName, countryCode, accountType, kycStatus, onboardingStatus, createdAt]
...
    Account:
      type: object
      properties:
        accountId:
          type: string
        userId:
          type: string
        currency:
          type: string
        balance:
          type: number
        availableBalance:
          type: number
        accountNumber:
          type: string
        sortCode:
          type: string
        iban:
          type: string
        swift:
          type: string
        accountType:
          type: string
          enum: [current, savings, vault, crypto]
        label:
          type: string
        color:
          type: string
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [accountId, userId, currency, balance, availableBalance, accountType, isDefault, createdAt]
    CreateAccountRequest:
      type: object
      properties:
        currency:
          type: string
        accountType:
          type: string
          enum: [current, savings, vault]
        label:
          type: string
      required: [currency, accountType]
    Transaction:
      type: object
      properties:
        transactionId:
          type: string
        accountId:
          type: string
        type:
          type: string
          enum: [credit, debit, transfer, fx_exchange, card_payment, atm, refund, topup]
        amount:
          type: number
        currency:
          type: string
        description:
          type: string
        merchantName:
          type: string
        merchantCategory:
          type: string
        status:
          type: string
          enum: [pending, completed, failed, reversed]
        createdAt:
          type: string
          format: date-time
        counterpartyName:
          type: string
        reference:
          type: string
      required: [transactionId, accountId, type, amount, currency, status, createdAt]
    TransactionPage:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
      required: [transactions, total, offset, limit]
    Card:
      type: object
      properties:
        cardId:
          type: string
        userId:
          type: string
        accountId:
          type: string
        cardType:
          type: string
          enum: [virtual, physical]
        cardNetwork:
          type: string
          enum: [visa, mastercard]
        lastFour:
          type: string
        expiryMonth:
          type: integer
        expiryYear:
          type: integer
        cardholderName:
          type: string
        status:
          type: string
          enum: [active, frozen, cancelled, pending_activation]
        spendingLimit:
          type: number
        currency:
          type: string
        color:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [cardId, userId, accountId, cardType, cardNetwork, lastFour, expiryMonth, expiryYear, cardholderName, status, currency, createdAt]
    IssueCardRequest:
      type: object
      properties:
        accountId:
          type: string
        cardType:
          type: string
          enum: [virtual, physical]
        cardNetwork:
          type: string
          enum: [visa, mastercard]
        spendingLimit:
          type: number
        label:
          type: string
      required: [accountId, cardType, cardNetwork]
    UpdateCardRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, frozen]
        spendingLimit:
          type: number
        label:
          type: string
...
    FxRates:
      type: object
      properties:
        base:
          type: string
        rates:
          type: object
          additionalProperties:
            type: number
        timestamp:
          type: string
          format: date-time
      required: [base, rates, timestamp]
    FxConvertRequest:
      type: object
      properties:
        fromCurrency:
          type: string
        toCurrency:
          type: string
        amount:
          type: number
        fromAccountId:
          type: string
        execute:
          type: boolean
          description: If true, actually execute the conversion
      required: [fromCurrency, toCurrency, amount]
    FxConversionResult:
      type: object
      properties:
        fromCurrency:
          type: string
        toCurrency:
          type: string
        fromAmount:
          type: number
        toAmount:
          type: number
        rate:
          type: number
        fee:
          type: number
        executed:
          type: boolean
        transactionId:
          type: string
      required: [fromCurrency, toCurrency, fromAmount, toAmount, rate, fee, executed]
    FxRatePoint:
      type: object
      properties:
        date:
          type: string
          format: date
        rate:
          type: number
      required: [date, rate]
...