# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Revolut-inspired reference fintech platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserAccountType } from "./userAccountType";
import type { UserKycStatus } from "./userKycStatus";
import type { UserOnboardingStatus } from "./userOnboardingStatus";

export interface User {
  id: string;
  email: string;
  phone?: string;
  firstName: string;
  lastName: string;
  countryCode: string;
  accountType: UserAccountType;
  kycStatus: UserKycStatus;
  onboardingStatus: UserOnboardingStatus;
  createdAt: Date;
}